+++ src/api/game.js
@@ -10,7 +10,7 @@ export const getGame = async ({queryKey}) => {
     },
     fields: {
       'node--game':
-        'title,body,genres,studios,genres,stores,releases_normalized,release_platforms,website,locations,publishers,sponsors,social_networks,languages,awards,credits,completeness,members,images,contextual_links,sources,article_links',
+        'title,body,genres,studios,genres,stores,releases_normalized,release_platforms,website,locations,cantons,publishers,sponsors,social_networks,languages,awards,credits,completeness,members,images,contextual_links,sources,article_links',
       studios: 'title',
       'node--studio': 'title,field_path',
       'node--people': 'title,field_path',
@@ -23,7 +23,7 @@ export const getGame = async ({queryKey}) => {
       'taxonomy_term--member': 'name',
     },
     include:
-      'studios,release_platforms,genres,locations,publishers,sponsors,languages,members,images',
+      'studios,release_platforms,genres,locations,cantons,publishers,sponsors,languages,members,images',
   });
 
   // console.log(`%cnode query: /node/game?${queryUrl}`, 'font-weight:bold;');