+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - add & fix missing translations strings for filters
 - fix footnote link visual glitch - Games-of-Switzerland/swissgamesgarden#9
 - fix about page links people - Games-of-Switzerland/swissgamesgarden#9
+- fix empty state games - Games-of-Switzerland/swissgamesgarden#28
 
 ## [0.1.0] - 2022-09-08
 ### Changed