+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - add no-index robots.txt on staging only - #102
+- add missing Stores links on Games detail - Games-of-Switzerland/swissgamesgarden#46
 
 ### Fixed
 - add & fix missing translations strings for filters