+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - add no-index robots.txt on staging only - #102
 - add missing Stores links on Games detail - Games-of-Switzerland/swissgamesgarden#46
+- allow link to be button - Games-of-Switzerland/swissgamesgarden#9
 
 ### Fixed
 - add & fix missing translations strings for filters