+++ CHANGELOG.md
@@ -6,6 +6,26 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2022-09-10
+### Changed
+- rework the deployment using Github environements
+- remove \[ & ] from the mailto subject of game udpate btn - Games-of-Switzerland/swissgamesgarden#88
+
+### Added
+- add no-index robots.txt on staging only - #102
+- add missing Stores links on Games detail - Games-of-Switzerland/swissgamesgarden#46
+- allow link to be button - Games-of-Switzerland/swissgamesgarden#9
+- add Cantons filters - Games-of-Switzerland/swissgamesgarden#51
+- add filter cantons - Games-of-Switzerland/swissgamesgarden#51
+- add cantons listing on game detail - Games-of-Switzerland/swissgamesgarden#51
+
+### Fixed
+- add & fix missing translations strings for filters
+- fix footnote link visual glitch - Games-of-Switzerland/swissgamesgarden#9
+- fix about page links people - Games-of-Switzerland/swissgamesgarden#9
+- fix empty state games - Games-of-Switzerland/swissgamesgarden#28
+- fix the members of games visual - Games-of-Switzerland/swissgamesgarden#98
+
 ## [0.1.0] - 2022-09-08
 ### Changed
 - update the hosts strategy to be compliante with new Docker container naming architecture
@@ -13,5 +33,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - update changelog following 'keep a changelog'
 
-[Unreleased]: https://github.com/Games-of-Switzerland/gos-website/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/Games-of-Switzerland/gos-website/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/Games-of-Switzerland/gos-website/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/Games-of-Switzerland/gos-website/releases/tag/0.1.0